+++ docs/theme/helpers/first.js
@@ -1,3 +1,0 @@
-module.exports = function(context, block) {
-  return block.fn(context[0]);
-};