+++ packages/ember-simple-auth/lib/ember-simple-auth/core.js
@@ -27,22 +27,6 @@ function shouldAuthorizeRequest(url) {
   return crossOriginWhitelist.indexOf(urlOrigin) > -1 || urlOrigin === documentOrigin;
 }
 
-function setupSession(store, container) {
-  var session = Session.create({ store: store, container: container });
-  var router  = container.lookup('router:main');
-  Ember.A([
-    'sessionAuthenticationSucceeded',
-    'sessionAuthenticationFailed',
-    'sessionInvalidationSucceeded',
-    'sessionInvalidationFailed'
-  ]).forEach(function(event) {
-    session.on(event, function(error) {
-      router.send(event, error);
-    });
-  });
-  return session;
-}
-
 var extensionInitializers = [];
 
 /**
@@ -142,7 +126,7 @@ var setup = function(container, application, options) {
 
   options.storeFactory = options.storeFactory || 'ember-simple-auth-session-store:local-storage';
   var store            = container.lookup(options.storeFactory);
-  var session          = setupSession(store, container);
+  var session          = Session.create({ store: store, container: container });
 
   container.register('ember-simple-auth-session:main', session, { instantiate: false });
   Ember.A(['controller', 'route']).forEach(function(component) {
@@ -156,6 +140,11 @@ var setup = function(container, application, options) {
       Ember.$.ajaxPrefilter(function(options, originalOptions, jqXHR) {
         if (shouldAuthorizeRequest(options.url)) {
           authorizer.authorize(jqXHR, options);
+        }
+      });
+      Ember.$(document).ajaxError(function(event, jqXHR, setting, exception) {
+        if (jqXHR.status === 401) {
+          session.trigger('authorizationFailed');
         }
       });
     }