+++ packages/ember-simple-auth/lib/ember-simple-auth/session.js
@@ -47,6 +47,50 @@ var global = (typeof window !== 'undefined') ? window : {},
 */
 var Session = Ember.ObjectProxy.extend(Ember.Evented, {
   /**
+    Triggered __whenever the session is successfully authenticated__. When the
+    application uses the mixin,
+    [Ember.SimpleAuth.ApplicationRouteMixin.actions#sessionAuthenticationSucceeded](#Ember-SimpleAuth-ApplicationRouteMixin-sessionAuthenticationSucceeded)
+    will be invoked whenever this event is triggered.
+
+    @event sessionAuthenticationSucceeded
+  */
+  /**
+    Triggered __whenever an attempt to authenticate the session fails__. When
+    the application uses the mixin,
+    [Ember.SimpleAuth.ApplicationRouteMixin.actions#sessionAuthenticationFailed](#Ember-SimpleAuth-ApplicationRouteMixin-sessionAuthenticationFailed)
+    will be invoked whenever this event is triggered.
+
+    @event sessionAuthenticationFailed
+    @param {Object} error The error object; this depends on the authenticator in use, see [Ember.SimpleAuth.Authenticators.Base#authenticate](#Ember-SimpleAuth-Authenticators-Base-authenticate)
+  */
+  /**
+    Triggered __whenever the session is successfully invalidated__. When the
+    application uses the mixin,
+    [Ember.SimpleAuth.ApplicationRouteMixin.actions#sessionInvalidationSucceeded](#Ember-SimpleAuth-ApplicationRouteMixin-sessionInvalidationSucceeded)
+    will be invoked whenever this event is triggered.
+
+    @event sessionInvalidationSucceeded
+  */
+  /**
+    Triggered __whenever an attempt to invalidate the session fails__. When the
+    application uses the mixin,
+    [Ember.SimpleAuth.ApplicationRouteMixin.actions#sessionInvalidationFailed](#Ember-SimpleAuth-ApplicationRouteMixin-sessionInvalidationFailed)
+    will be invoked whenever this event is triggered.
+
+    @event sessionInvalidationFailed
+    @param {Object} error The error object; this depends on the authenticator in use, see [Ember.SimpleAuth.Authenticators.Base#invalidate](#Ember-SimpleAuth-Authenticators-Base-invalidate)
+  */
+  /**
+    Triggered __whenever the server rejects the authorization information
+    passed with a request and responds with status 401__. When the application
+    uses the mixin,
+    [Ember.SimpleAuth.ApplicationRouteMixin.actions#authorizationFailed](#Ember-SimpleAuth-ApplicationRouteMixin-authorizationFailed)
+    will be invoked whenever this event is triggered.
+
+    @event authorizationFailed
+  */
+
+  /**
     The authenticator factory used to authenticate the session. This is only
     set when the session is currently authenticated.
 
@@ -150,7 +194,7 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
     return new Ember.RSVP.Promise(function(resolve, reject) {
       var authenticator = _this.container.lookup(_this.authenticatorFactory);
       authenticator.invalidate(_this.content).then(function() {
-        authenticator.off('updated');
+        authenticator.off('sessionDataUpdated');
         _this.clear(true);
         resolve();
       }, function(error) {
@@ -191,6 +235,7 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
   */
   setup: function(authenticatorFactory, content, trigger) {
     trigger = !!trigger && !this.get('isAuthenticated');
+    this.beginPropertyChanges();
     this.setProperties({
       isAuthenticated:      true,
       authenticatorFactory: authenticatorFactory,
@@ -199,6 +244,7 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
     this.bindToAuthenticatorEvents();
     var data = Ember.$.extend({ authenticatorFactory: authenticatorFactory }, this.content);
     this.store.replace(data);
+    this.endPropertyChanges();
     if (trigger) {
       this.trigger('sessionAuthenticationSucceeded');
     }
@@ -210,12 +256,14 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
   */
   clear: function(trigger) {
     trigger = !!trigger && this.get('isAuthenticated');
+    this.beginPropertyChanges();
     this.setProperties({
       isAuthenticated:      false,
       authenticatorFactory: null,
       content:              {}
     });
     this.store.clear();
+    this.endPropertyChanges();
     if (trigger) {
       this.trigger('sessionInvalidationSucceeded');
     }
@@ -228,12 +276,12 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
   bindToAuthenticatorEvents: function() {
     var _this = this;
     var authenticator = this.container.lookup(this.authenticatorFactory);
-    authenticator.off('updated');
-    authenticator.off('invalidated');
-    authenticator.on('updated', function(content) {
+    authenticator.off('sessionDataUpdated');
+    authenticator.off('sessionDataInvalidated');
+    authenticator.on('sessionDataUpdated', function(content) {
       _this.setup(_this.authenticatorFactory, content);
     });
-    authenticator.on('invalidated', function(content) {
+    authenticator.on('sessionDataInvalidated', function(content) {
       _this.clear(true);
     });
   },
@@ -244,7 +292,7 @@ var Session = Ember.ObjectProxy.extend(Ember.Evented, {
   */
   bindToStoreEvents: function() {
     var _this = this;
-    this.store.on('updated', function(content) {
+    this.store.on('sessionDataUpdated', function(content) {
       var authenticatorFactory = content.authenticatorFactory;
       if (!!authenticatorFactory) {
         delete content.authenticatorFactory;