+++ packages/ember-simple-auth/tests/core_test.js
@@ -183,6 +183,56 @@ describe('setup', function() {
       });
     });
 
+    describe('the AJAX error handler', function() {
+      beforeEach(function() {
+        this.xhr                = sinon.useFakeXMLHttpRequest();
+        this.server             = sinon.fakeServer.create();
+        this.server.autoRespond = true;
+        sinon.spy(this.container, 'register');
+        setup(this.container, this.application, { authorizerFactory: 'authorizerFactory' });
+        var spyCall  = this.container.register.getCall(2);
+        this.session = spyCall.args[1];
+      });
+
+      describe("when the request's status is 401", function() {
+        beforeEach(function() {
+          this.server.respondWith('GET', '/data', [401, {}, '']);
+        });
+
+        it("triggers the session's authorizationFailed event", function(done) {
+          var triggered = false;
+          this.session.one('authorizationFailed', function() { triggered = true; });
+          Ember.$.get('/data');
+
+          Ember.run.later(function() {
+            expect(triggered).to.be.true;
+            done();
+          }, 100);
+        });
+      });
+
+      describe("when the request's status is not 401", function() {
+        beforeEach(function() {
+          this.server.respondWith('GET', '/data', [500, {}, '']);
+        });
+
+        it("does not trigger the session's authorizationFailed event", function(done) {
+          var triggered = false;
+          this.session.one('authorizationFailed', function() { triggered = true; });
+          Ember.$.get('/data');
+
+          Ember.run.later(function() {
+            expect(triggered).to.be.false;
+            done();
+          }, 100);
+        });
+      });
+
+      afterEach(function() {
+        this.xhr.restore();
+      });
+    });
+
     afterEach(function() {
       Ember.$.ajaxPrefilter.restore();
     });
@@ -194,52 +244,6 @@ describe('setup', function() {
       setup(this.container, this.application);
 
       expect(Ember.$.ajaxPrefilter).to.not.have.been.called;
-    });
-  });
-
-  describe('session events', function() {
-    beforeEach(function() {
-      sinon.spy(this.router, 'send');
-      sinon.spy(this.container, 'register');
-      setup(this.container, this.application);
-      var spyCall  = this.container.register.getCall(2);
-      this.session = spyCall.args[1];
-    });
-
-    it('forwards the "sessionAuthenticationSucceeded" event to the router', function(done) {
-      this.session.trigger('sessionAuthenticationSucceeded');
-
-      Ember.run.next(this, function() {
-        expect(this.router.send).to.have.been.calledWith('sessionAuthenticationSucceeded');
-        done();
-      });
-    });
-
-    it('forwards the "sessionAuthenticationFailed" event to the router', function(done) {
-      this.session.trigger('sessionAuthenticationFailed', 'error');
-
-      Ember.run.next(this, function() {
-        expect(this.router.send).to.have.been.calledWith('sessionAuthenticationFailed', 'error');
-        done();
-      });
-    });
-
-    it('forwards the "sessionInvalidationSucceeded" event to the router', function(done) {
-      this.session.trigger('sessionInvalidationSucceeded');
-
-      Ember.run.next(this, function() {
-        expect(this.router.send).to.have.been.calledWith('sessionInvalidationSucceeded');
-        done();
-      });
-    });
-
-    it('forwards the "sessionInvalidationFailed" to the router', function(done) {
-      this.session.trigger('sessionInvalidationFailed', 'error');
-
-      Ember.run.next(this, function() {
-        expect(this.router.send).to.have.been.calledWith('sessionInvalidationFailed', 'error');
-        done();
-      });
     });
   });