+++ packages/ember-simple-auth/tests/mixins/authentication_controller_mixin_test.js
@@ -17,5 +17,12 @@ describe('AuthenticationControllerMixin', function() {
 
       expect(this.session.authenticate).to.have.been.calledWith('authenticatorFactory', { some: 'options' });
     });
+
+    it('returns the prmoise returned by the session', function() {
+      var promise = new Ember.RSVP.Promise(function() {});
+      sinon.stub(this.session, 'authenticate').returns(promise);
+
+      expect(this.controller._actions.authenticate.apply(this.controller, [{ some: 'options' }])).to.eq(promise);
+    });
   });
 });