+++ packages/ember-simple-auth/tests/session_test.js
@@ -17,13 +17,11 @@ describe('Session', function() {
       });
 
       it('updates its content', function(done) {
-        Ember.run.next(this, function() {
-          this.authenticator.trigger('updated', { some: 'other property' });
+        this.authenticator.trigger('sessionDataUpdated', { some: 'other property' });
 
-          Ember.run.next(this, function() {
-            expect(this.session.get('content')).to.eql({ some: 'other property' });
-            done();
-          });
+        Ember.run.next(this, function() {
+          expect(this.session.get('content')).to.eql({ some: 'other property' });
+          done();
         });
       });
     });
@@ -34,49 +32,40 @@ describe('Session', function() {
       });
 
       it('is not authenticated', function(done) {
-        Ember.run.next(this, function() {
-          this.authenticator.trigger('invalidated', { some: 'other property' });
+        this.authenticator.trigger('sessionDataInvalidated', { some: 'other property' });
 
-          Ember.run.next(this, function() {
-            expect(this.session.get('isAuthenticated')).to.be.false;
-            done();
-          });
+        Ember.run.next(this, function() {
+          expect(this.session.get('isAuthenticated')).to.be.false;
+          done();
         });
       });
 
       it('clears its content', function(done) {
-        Ember.run.next(this, function() {
-          this.authenticator.trigger('invalidated', { some: 'other property' });
+        this.authenticator.trigger('sessionDataInvalidated', { some: 'other property' });
 
-          Ember.run.next(this, function() {
-            expect(this.session.get('content')).to.eql({});
-            done();
-          });
+        Ember.run.next(this, function() {
+          expect(this.session.get('content')).to.eql({});
+          done();
         });
       });
 
       it('clears the store', function(done) {
-        Ember.run.next(this, function() {
-          this.authenticator.trigger('invalidated', { some: 'other property' });
+        this.authenticator.trigger('sessionDataInvalidated', { some: 'other property' });
 
-          Ember.run.next(this, function() {
-            expect(this.store.restore()).to.eql({});
-            done();
-          });
+        Ember.run.next(this, function() {
+          expect(this.store.restore()).to.eql({});
+          done();
         });
       });
 
       it('triggers the "sessionInvalidationSucceeded" event', function(done) {
         var triggered = false;
         this.session.one('sessionInvalidationSucceeded', function() { triggered = true; });
+        this.authenticator.trigger('sessionDataInvalidated', { some: 'other property' });
 
         Ember.run.next(this, function() {
-          this.authenticator.trigger('invalidated', { some: 'other property' });
-
-          Ember.run.next(this, function() {
-            expect(triggered).to.be.true;
-            done();
-          });
+          expect(triggered).to.be.true;
+          done();
         });
       });
     });
@@ -525,7 +514,7 @@ describe('Session', function() {
         });
 
         it('is authenticated', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             expect(this.session.get('isAuthenticated')).to.be.true;
@@ -534,7 +523,7 @@ describe('Session', function() {
         });
 
         it('sets its content to the data the authenticator resolves with', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             var properties = this.store.restore();
@@ -546,7 +535,7 @@ describe('Session', function() {
         });
 
         it('persists its content in the store', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             var properties = this.store.restore();
@@ -558,7 +547,7 @@ describe('Session', function() {
         });
 
         it('persists its content in the store', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             var properties = this.store.restore();
@@ -577,7 +566,7 @@ describe('Session', function() {
           it('does not trigger the "sessionAuthenticationSucceeded" event', function(done) {
             var triggered = false;
             this.session.one('sessionAuthenticationSucceeded', function() { triggered = true; });
-            this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+            this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
             Ember.run.next(this, function() {
               expect(triggered).to.be.false;
@@ -594,7 +583,7 @@ describe('Session', function() {
           it('triggers the "sessionAuthenticationSucceeded" event', function(done) {
             var triggered = false;
             this.session.one('sessionAuthenticationSucceeded', function() { triggered = true; });
-            this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+            this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
             Ember.run.next(this, function() {
               expect(triggered).to.be.true;
@@ -610,7 +599,7 @@ describe('Session', function() {
         });
 
         it('is not authenticated', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             expect(this.session.get('isAuthenticated')).to.be.false;
@@ -619,7 +608,7 @@ describe('Session', function() {
         });
 
         it('clears its content', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             expect(this.session.get('content')).to.eql({});
@@ -628,7 +617,7 @@ describe('Session', function() {
         });
 
         it('clears the store', function(done) {
-          this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
           Ember.run.next(this, function() {
             expect(this.store.restore()).to.eql({});
@@ -644,7 +633,7 @@ describe('Session', function() {
           it('triggers the "sessionInvalidationSucceeded" event', function(done) {
             var triggered = false;
             this.session.one('sessionInvalidationSucceeded', function() { triggered = true; });
-            this.store.trigger('updated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
+            this.store.trigger('sessionDataUpdated', { some: 'other property', authenticatorFactory: 'authenticatorFactory' });
 
             Ember.run.next(this, function() {
               expect(triggered).to.be.true;
@@ -673,7 +662,7 @@ describe('Session', function() {
 
     describe('when there is no authenticator factory in the store', function() {
       it('is not authenticated', function(done) {
-        this.store.trigger('updated', { some: 'other property' });
+        this.store.trigger('sessionDataUpdated', { some: 'other property' });
 
         Ember.run.next(this, function() {
           expect(this.session.get('isAuthenticated')).to.be.false;
@@ -682,7 +671,7 @@ describe('Session', function() {
       });
 
       it('clears its content', function(done) {
-        this.store.trigger('updated', { some: 'other property' });
+        this.store.trigger('sessionDataUpdated', { some: 'other property' });
 
         Ember.run.next(this, function() {
           expect(this.session.get('content')).to.eql({});
@@ -691,7 +680,7 @@ describe('Session', function() {
       });
 
       it('clears the store', function(done) {
-        this.store.trigger('updated', { some: 'other property' });
+        this.store.trigger('sessionDataUpdated', { some: 'other property' });
 
         Ember.run.next(this, function() {
           expect(this.store.restore()).to.eql({});
@@ -707,7 +696,7 @@ describe('Session', function() {
         it('triggers the "sessionInvalidationSucceeded" event', function(done) {
           var triggered = false;
           this.session.one('sessionInvalidationSucceeded', function() { triggered = true; });
-          this.store.trigger('updated', { some: 'other property' });
+          this.store.trigger('sessionDataUpdated', { some: 'other property' });
 
           Ember.run.next(this, function() {
             expect(triggered).to.be.true;