+++ test/lib/dependencies_loader.js
@@ -4,9 +4,8 @@
     return (window.location.search.match(regexp) || [])[1];
   }
 
-  var jQueryVersion     = parseVersion('jQuery') || '2.1.0';
-  var handlebarsVersion = parseVersion('handlebars') || '1.3.0';
-  var emberVersion      = parseVersion('ember') || '1.4.0';
+  var jQueryVersion     = parseVersion('jQuery') || '2.1.1';
+  var emberVersion      = parseVersion('ember') || '1.5.1';
 
   document.write('<script src="vendor/jquery-' + jQueryVersion + '.js"></script>');
   document.write('<script src="vendor/handlebars.js"></script>');